Some wording changes
I changed some wording in the instructions that mentioned "Two"
assignments. I removed that because in this alternate version, there is
no Data Journal. Since I did this in wordpad, I need the document to be
reformated. Please keep in mind to follow the page breaks in the
original pdf!